+++ src/_components/common/Paginator/hooks/useDeviceWidth.ts
