+++ src/_constants/questionCategory.ts
@@ -1,4 +1,4 @@
-export const CATEGORY = {
+export const QUESTION_CATEGORY = {
   전체: 'all',
   프론트엔드: 'frontend',
   백엔드: 'backend',
